[backend_setup_make_migration] - modify the codes 2

+++ database/migrations/2024_05_03_122855_create_reservations_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('area_id');
             $table->date('date');
-            $table->integer('fee_log');
+            $table->double('fee_log', 5, 2);
             $table->timestamps();
             $table->softDeletes();
